add code edit services

+++ database/migrations/2023_10_15_105217_create_service_table.php
@@ -15,14 +15,14 @@ return new class extends Migration
         Schema::create('services', function (Blueprint $table) {
             $table->id();
             $table->unsignedBigInteger("user_id");
-            $table->decimal("electricityBill", 10, 3)->nullable();//tiền 1 số điện
-            $table->decimal("waterBill", 10, 3)->nullable();//tiền 1 khối nước
-            $table->decimal("wifiFee", 10, 3)->nullable();//tiền wifi
-            $table->decimal("cleaningFee", 10, 3)->nullable();//dọn dẹp
-            $table->decimal("parkingFee", 10, 3)->nullable();// tiền để xe
-            $table->decimal("fine", 10, 3)->nullable();//tiền phạt
-            $table->decimal("other_fees", 10, 3)->nullable();// khoản khác
-            $table->decimal("sumServices", 10, 3);//tổng tiền dịch vụ
+            $table->float("electricityBill", 10, 3)->nullable();//tiền 1 số điện
+            $table->float("waterBill", 10, 3)->nullable();//tiền 1 khối nước
+            $table->float("wifiFee", 10, 3)->nullable();//tiền wifi
+            $table->float("cleaningFee", 10, 3)->nullable();//dọn dẹp
+            $table->float("parkingFee", 10, 3)->nullable();// tiền để xe
+            $table->float("fine", 10, 3)->nullable();//tiền phạt
+            $table->float("other_fees", 10, 3)->nullable();// khoản khác
+            $table->float("sumServices", 10, 3);//tổng tiền dịch vụ
             $table->timestamps();
             // Tạo khóa ngoại cho user_id
             $table->foreign('user_id')->references('id')->on('users');
